refactor(store): decomposition ngxs config
