+++ packages/store/src/decorators/select.ts
@@ -2,7 +2,8 @@ import { Injectable } from '@angular/core';
 
 import { Store } from '../store';
 import { propGetter, removeDollarAtTheEnd } from '../internal/internals';
-import { META_KEY, NgxsConfig } from '../symbols';
+import { META_KEY } from '../symbols';
+import { NgxsConfig } from '../internal/config/ngxs-config';
 
 /**
  * Allows the select decorator to get access to the DI store.
